// frontend/src/app/core/api/model/requisitionCreateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RequisitionCreateDto { 
    title?: string;
    description?: string;
    amount?: number;
}